Add support for firecracker VM snapshot-restore
In order to support firecracker snapshot-restore we need to make the
following changes:
* add a new CreateSnapshot request;
* add new parameters for snapshot loading to the CreateVM request
(following the firecracker Golang SDK design, snapshot loading is
essentially creating a VM with additional snapshot options).

When a VM is created by loading a snapshot, it's container snapshot drives
are already mounted, so the drive mount stub creation and mounting is
skipped.

Also add a NetNS (network namespace) parameter to the CreateVM request,
since the current NetNS parameter nested in the JailerConfig parameter
requires specifying the entire jailer configuration, which is not
convenient.

Signed-off-by: Georgiy Lebedev <lebedev.gk@phystech.edu>